chore: Add Bash script for automated file deletion
Introduce a Bash script that automates the deletion of files in a specified target folder older than 3 days. The script logs the deleted files along with their paths to a log file located in the user's Documents folder.

Note: Users need to customize the 'target_folder' variable with the path to their intended folder.